Add card view and routes for info page
This commit uses the normal card view to the home screen, allowing users to see their cards in a list in place of the previous image. It also adds routes for the info page, so users can navigate to it from the home screen. And i've removed the search bar to be added in upcoming releases